# handlers/auth.go
package handlers

import (
	"fmt"
	"github.com/gin-gonic/gin"
	"github.com/happyphper/gin-swagger-demo/response"
	"net/http"
	"strings"
	"time"
)

type LoginForm struct {
	// 账号
	Username string `json:"username" validate:"required,min=2" example:"happyphper"`
	// 密码
	Password string `json:"password" validate:"required,min=6" example:"123123123"`
}

// Login
// @Summary 登录
// @Description 登录接口，适用于账号、密码登录
// @Tags Auth
// @Accept json
// @Produce json
// @Param username body handlers.LoginForm false "JSON"
// @Success 200 {object} response.Response{code=int,data=response.TokenModel} "desc"
// @Success 200 {object} response.Response{code=int,data=response.TokenModel} "desc"
// @Router /login [post]
func Login(ctx *gin.Context) {
	var form LoginForm
	if err := ctx.ShouldBindJSON(&form); err != nil {
		ctx.JSON(http.StatusOK, response.ValidationRes(err.Error()))
		return
	}

	if form.Username != "happyphper" || form.Password != "123123123" {
		ctx.JSON(http.StatusOK, response.AuthErrorRes())
		return
	}

	fmt.Println(form)

	tokenModel := response.TokenModel{
		Token:     "110123123",
		ExpiresIn: time.Now().Add(time.Hour * 24 * 15).Unix(),
	}

	ctx.JSON(http.StatusOK, tokenModel.ToResponse())
}

type RegisterForm struct {
	// 账号
	Username string `json:"username" validate:"required,min=2" example:"18010021234"`
	// 密码
	Password string `json:"password" validate:"required,min=6" example:"123123123"`
}
// Register
// @Summary 注册
// @Description 适用于新账号注册
// @Tags Auth
// @Accept json
// @Produce json
// @Param username body handlers.RegisterForm false "JSON"
// @Success 200 {object} response.Response{code=int,data=response.TokenModel} "desc"
// @Success 200 {object} response.Response{code=int,data=response.TokenModel} "desc"
// @Router /register [post]
func Register(ctx *gin.Context) {
	var form LoginForm
	if err := ctx.BindJSON(&form); err != nil {
		ctx.JSON(http.StatusOK, response.ValidationRes(err.Error()))
		return
	}

	if form.Username == "happyphper" {
		ctx.JSON(http.StatusOK, response.UsernameExistsRes())
		return
	}

	tokenModel := response.TokenModel{
		Token:        "110123123",
		ExpiresIn: time.Now().Add(time.Hour * 24 * 15).Unix(),
	}

	ctx.JSON(http.StatusOK, tokenModel.ToResponse())
}


// Me
// @Summary 我的
// @Description 个人信息
// @Tags Auth
// @Accept json
// @Produce json
// @Security BearerTokenAuth
// @Success 200 {object} response.Response{code=int,data=models.User} "desc"
// @Router /me [get]
func Me(ctx *gin.Context) {
	// TODO 封装为中间件
	auth := ctx.GetHeader("Authorization")
	arr := strings.SplitN(auth, " ", 2)
	if arr[0] != "Bearer" || arr[1] != "123" {
		ctx.JSON(http.StatusOK, response.TokenExpiredRes())
		return
	}

	var form LoginForm
	if err := ctx.ShouldBindJSON(&form); err != nil {
		ctx.JSON(http.StatusOK, response.ValidationRes(err.Error()))
		return
	}

	if form.Username != "happyphper" || form.Password != "123123123" {
		ctx.JSON(http.StatusOK, response.AuthErrorRes())
		return
	}

	fmt.Println(form)

	tokenModel := response.TokenModel{
		Token:     "110123123",
		ExpiresIn: time.Now().Add(time.Hour * 24 * 15).Unix(),
	}

	ctx.JSON(http.StatusOK, tokenModel.ToResponse())
}